Added story and fix prop call
